feat: add expect on test insertionWithFile
