+++ api/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node app.js",
     "dev": "nodemon -e .js app.js",
-    "test": "nodemon -w test/ -e .js --exec 'mocha' --timeout 30000",
+    "test": "nodemon -w test/ -e .js --exec 'mocha' --timeout 10000",
     "lint": "eslint --ext .js"
   },
   "devDependencies": {
