+++ api/test/insertionWithFile.js
@@ -3,11 +3,6 @@ const { expect } = require('chai');
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 
-const api = require('../app');
-const fakeUnpaywall = require('../../fakeUnpaywall/app');
-
-const client = require('../lib/client');
-
 const indexUnpawall = require('../index/unpaywall.json');
 const indexTask = require('../index/task.json');
 
@@ -16,71 +11,52 @@ const {
   deleteIndex,
   countDocuments,
   isTaskEnd,
+  getTask,
+} = require('./utils/elastic');
+
+const {
+  getLatestReport,
+} = require('./utils/report');
+
+const {
+  initializeDate,
   deleteFile,
   downloadFile,
-  getTask,
-  getLatestReport,
-} = require('./utils');
-const { logger } = require('../lib/logger');
+} = require('./utils/file');
 
-chai.should();
+const {
+  ping,
+} = require('./utils/ping');
+
 chai.use(chaiHttp);
 
-describe('test insertion with a file already installed in ez-unpaywall', () => {
-  const ezunpaywallURL = 'http://localhost:8080';
-  const fakeUnpaywallURL = 'http://localhost:12000';
+const ezunpaywallURL = 'http://localhost:8080';
 
+describe('Test: insert the content of a file already installed on ezunpaywall', () => {
   before(async () => {
-    // wait ezunpaywall
-    let res1;
-    while (res1?.body?.data !== 'pong') {
-      try {
-        res1 = await chai.request(ezunpaywallURL).get('/ping');
-      } catch (err) {
-        logger.error(`ezunpaywall ping : ${err}`);
-      }
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-    }
-    // wait fakeUnpaywall
-    let res2;
-    while (res2?.body?.data !== 'pong') {
-      try {
-        res2 = await chai.request(fakeUnpaywallURL).get('/ping');
-      } catch (err) {
-        logger.error(`fakeUnpaywall ping : ${err}`);
-      }
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-    }
-    // wait elastic started
-    let res3;
-    while (res3?.statusCode !== 200) {
-      try {
-        res3 = await client.ping();
-      } catch (err) {
-        logger.error(`elastic ping : ${err}`);
-      }
-      await new Promise((resolve) => setTimeout(resolve, 1000));
-    }
+    await ping();
+    initializeDate();
     await downloadFile('fake1.jsonl.gz');
   });
 
-  describe('/update/fake1.jsonl.gz insert a file already installed', () => {
+  describe('Do a classic insertion of a file already installed', () => {
     before(async () => {
       await createIndex('task', indexTask);
       await createIndex('unpaywall', indexUnpawall);
     });
     // test return message
-    it('should return the process start', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return the process start', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl.gz')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
-      response.should.have.status(200);
-      response.body.should.have.property('message').equal('start upsert with fake1.jsonl.gz');
+
+      expect(res).have.status(200);
+      expect(res.body.message).be.equal('start upsert with fake1.jsonl.gz');
     });
 
     // test insertion
-    it('should insert 50 datas', async () => {
+    it('Should insert 50 datas', async () => {
       let taskEnd;
       while (!taskEnd) {
         await new Promise((resolve) => setTimeout(resolve, 1000));
@@ -91,41 +67,41 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
 
     // test task
-    it('should get task with all informations', async () => {
+    it('Should get task with all informations from the insertion', async () => {
       const task = await getTask();
 
-      task.should.have.property('done').equal(true);
-      task.should.have.property('currentTask').equal('end');
-      task.should.have.property('steps');
-      task.should.have.property('createdAt');
-      task.should.have.property('endAt');
-      task.should.have.property('took');
+      expect(task).have.property('done').equal(true);
+      expect(task).have.property('currentTask').equal('end');
+      expect(task).have.property('steps');
+      expect(task).have.property('createdAt');
+      expect(task).have.property('endAt');
+      expect(task).have.property('took');
 
-      task.steps[0].should.have.property('task').equal('insert');
-      task.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      task.steps[0].should.have.property('percent').equal(100);
-      task.steps[0].should.have.property('lineRead').equal(50);
-      task.steps[0].should.have.property('took');
-      task.steps[0].should.have.property('status').equal('success');
+      expect(task.steps[0]).have.property('task').equal('insert');
+      expect(task.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(task.steps[0]).have.property('percent').equal(100);
+      expect(task.steps[0]).have.property('lineRead').equal(50);
+      expect(task.steps[0]).have.property('took');
+      expect(task.steps[0]).have.property('status').equal('success');
     });
 
     // test Report
-    it('should get report with all informations', async () => {
+    it('Should get report with all informations from the insertion', async () => {
       const report = await getLatestReport();
 
-      report.should.have.property('done').equal(true);
-      report.should.have.property('currentTask').equal('end');
-      report.should.have.property('steps');
-      report.should.have.property('createdAt');
-      report.should.have.property('endAt');
-      report.should.have.property('took');
+      expect(report).have.property('done').equal(true);
+      expect(report).have.property('currentTask').equal('end');
+      expect(report).have.property('steps');
+      expect(report).have.property('createdAt');
+      expect(report).have.property('endAt');
+      expect(report).have.property('took');
 
-      report.steps[0].should.have.property('task').equal('insert');
-      report.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      report.steps[0].should.have.property('percent').equal(100);
-      report.steps[0].should.have.property('lineRead').equal(50);
-      report.steps[0].should.have.property('took');
-      report.steps[0].should.have.property('status').equal('success');
+      expect(report.steps[0]).have.property('task').equal('insert');
+      expect(report.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(report.steps[0]).have.property('percent').equal(100);
+      expect(report.steps[0]).have.property('lineRead').equal(50);
+      expect(report.steps[0]).have.property('took');
+      expect(report.steps[0]).have.property('status').equal('success');
     });
 
     after(async () => {
@@ -134,20 +110,20 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
   });
 
-  describe('/update/fake1.jsonl.gz?limit=10 insert a file already installed with limit 10', () => {
+  describe('Do a classic insertion of a file already installed with parameter limit=10', () => {
     // test return message
-    it('should return the process start', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return the process start', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl.gz?limit=10')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
 
-      response.should.have.status(200);
-      response.body.should.have.property('message').equal('start upsert with fake1.jsonl.gz');
+      expect(res).have.status(200);
+      expect(res.body.message).be.equal('start upsert with fake1.jsonl.gz');
     });
 
     // test insertion
-    it('should insert 10 datas', async () => {
+    it('Should insert 10 data', async () => {
       let taskEnd;
       while (!taskEnd) {
         await new Promise((resolve) => setTimeout(resolve, 1000));
@@ -158,41 +134,41 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
 
     // test task
-    it('should get task with all informations', async () => {
+    it('Should get task with all informations from the insertion', async () => {
       const task = await getTask();
 
-      task.should.have.property('done').equal(true);
-      task.should.have.property('currentTask').equal('end');
-      task.should.have.property('steps');
-      task.should.have.property('createdAt');
-      task.should.have.property('endAt');
-      task.should.have.property('took');
+      expect(task).have.property('done').equal(true);
+      expect(task).have.property('currentTask').equal('end');
+      expect(task).have.property('steps');
+      expect(task).have.property('createdAt');
+      expect(task).have.property('endAt');
+      expect(task).have.property('took');
 
-      task.steps[0].should.have.property('task').equal('insert');
-      task.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      task.steps[0].should.have.property('percent').equal(100);
-      task.steps[0].should.have.property('lineRead').equal(10);
-      task.steps[0].should.have.property('took');
-      task.steps[0].should.have.property('status').equal('success');
+      expect(task.steps[0]).have.property('task').equal('insert');
+      expect(task.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(task.steps[0]).have.property('percent').equal(100);
+      expect(task.steps[0]).have.property('lineRead').equal(10);
+      expect(task.steps[0]).have.property('took');
+      expect(task.steps[0]).have.property('status').equal('success');
     });
 
     // test report
-    it('should get report with all informations', async () => {
+    it('Should get report with all informations from the insertion', async () => {
       const report = await getLatestReport();
 
-      report.should.have.property('done').equal(true);
-      report.should.have.property('currentTask').equal('end');
-      report.should.have.property('steps');
-      report.should.have.property('createdAt');
-      report.should.have.property('endAt');
-      report.should.have.property('took');
+      expect(report).have.property('done').equal(true);
+      expect(report).have.property('currentTask').equal('end');
+      expect(report).have.property('steps');
+      expect(report).have.property('createdAt');
+      expect(report).have.property('endAt');
+      expect(report).have.property('took');
 
-      report.steps[0].should.have.property('task').equal('insert');
-      report.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      report.steps[0].should.have.property('percent').equal(100);
-      report.steps[0].should.have.property('lineRead').equal(10);
-      report.steps[0].should.have.property('took');
-      report.steps[0].should.have.property('status').equal('success');
+      expect(report.steps[0]).have.property('task').equal('insert');
+      expect(report.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(report.steps[0]).have.property('percent').equal(100);
+      expect(report.steps[0]).have.property('lineRead').equal(10);
+      expect(report.steps[0]).have.property('took');
+      expect(report.steps[0]).have.property('status').equal('success');
     });
 
     after(async () => {
@@ -201,23 +177,24 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
   });
 
-  describe('/update/fake1.jsonl.gz?offset=40 insert a file already installed with offset 40', () => {
+  describe('Do a classic insertion of a file already installed with parameter offset=40', () => {
     before(async () => {
       await createIndex('task', indexTask);
       await createIndex('unpaywall', indexUnpawall);
     });
     // test return message
     it('should return the process start', async () => {
-      const response = await chai.request(ezunpaywallURL)
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl.gz?offset=40')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
-      response.should.have.status(200);
-      response.body.should.have.property('message').equal('start upsert with fake1.jsonl.gz');
+
+      expect(res).have.status(200);
+      expect(res.body.message).be.equal('start upsert with fake1.jsonl.gz');
     });
 
     // test insertion
-    it('should insert 10 datas', async () => {
+    it('Should insert 10 datas', async () => {
       let taskEnd;
       while (!taskEnd) {
         await new Promise((resolve) => setTimeout(resolve, 1000));
@@ -228,41 +205,41 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
 
     // test task
-    it('should get task with all informations', async () => {
+    it('Should get task with all informations from the insertion', async () => {
       const task = await getTask();
 
-      task.should.have.property('done').equal(true);
-      task.should.have.property('currentTask').equal('end');
-      task.should.have.property('steps');
-      task.should.have.property('createdAt');
-      task.should.have.property('endAt');
-      task.should.have.property('took');
+      expect(task).have.property('done').equal(true);
+      expect(task).have.property('currentTask').equal('end');
+      expect(task).have.property('steps');
+      expect(task).have.property('createdAt');
+      expect(task).have.property('endAt');
+      expect(task).have.property('took');
 
-      task.steps[0].should.have.property('task').equal('insert');
-      task.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      task.steps[0].should.have.property('percent').equal(100);
-      task.steps[0].should.have.property('lineRead').equal(50);
-      task.steps[0].should.have.property('took');
-      task.steps[0].should.have.property('status').equal('success');
+      expect(task.steps[0]).have.property('task').equal('insert');
+      expect(task.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(task.steps[0]).have.property('percent').equal(100);
+      expect(task.steps[0]).have.property('lineRead').equal(50);
+      expect(task.steps[0]).have.property('took');
+      expect(task.steps[0]).have.property('status').equal('success');
     });
 
     // test report
-    it('should get task with all informations', async () => {
+    it('Should get task with all informations from the insertion', async () => {
       const report = await getLatestReport();
 
-      report.should.have.property('done').equal(true);
-      report.should.have.property('currentTask').equal('end');
-      report.should.have.property('steps');
-      report.should.have.property('createdAt');
-      report.should.have.property('endAt');
-      report.should.have.property('took');
+      expect(report).have.property('done').equal(true);
+      expect(report).have.property('currentTask').equal('end');
+      expect(report).have.property('steps');
+      expect(report).have.property('createdAt');
+      expect(report).have.property('endAt');
+      expect(report).have.property('took');
 
-      report.steps[0].should.have.property('task').equal('insert');
-      report.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      report.steps[0].should.have.property('percent').equal(100);
-      report.steps[0].should.have.property('lineRead').equal(50);
-      report.steps[0].should.have.property('took');
-      report.steps[0].should.have.property('status').equal('success');
+      expect(report.steps[0]).have.property('task').equal('insert');
+      expect(report.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(report.steps[0]).have.property('percent').equal(100);
+      expect(report.steps[0]).have.property('lineRead').equal(50);
+      expect(report.steps[0]).have.property('took');
+      expect(report.steps[0]).have.property('status').equal('success');
     });
 
     after(async () => {
@@ -271,23 +248,24 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
   });
 
-  describe('/update/fake1.jsonl.gz?offset=10&limit=20 insert a file already installed with limit=10 and offset=20', () => {
+  describe('Do a classic insertion of a file already installed with parameter offset=10 and limit=20', () => {
     before(async () => {
       await createIndex('task', indexTask);
       await createIndex('unpaywall', indexUnpawall);
     });
     // test return message
-    it('should return the process start', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return the process start', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl.gz?offset=10&limit=20')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
-      response.should.have.status(200);
-      response.body.should.have.property('message').equal('start upsert with fake1.jsonl.gz');
+
+      expect(res).have.status(200);
+      expect(res.body.message).be.equal('start upsert with fake1.jsonl.gz');
     });
 
     // test insertion
-    it('should insert 10 datas', async () => {
+    it('Should insert 10 data', async () => {
       let taskEnd;
       while (!taskEnd) {
         await new Promise((resolve) => setTimeout(resolve, 1000));
@@ -298,41 +276,41 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
 
     // test task
-    it('should get task with all informations', async () => {
+    it('Should get task with all informations from the insertion', async () => {
       const task = await getTask();
 
-      task.should.have.property('done').equal(true);
-      task.should.have.property('currentTask').equal('end');
-      task.should.have.property('steps');
-      task.should.have.property('createdAt');
-      task.should.have.property('endAt');
-      task.should.have.property('took');
+      expect(task).have.property('done').equal(true);
+      expect(task).have.property('currentTask').equal('end');
+      expect(task).have.property('steps');
+      expect(task).have.property('createdAt');
+      expect(task).have.property('endAt');
+      expect(task).have.property('took');
 
-      task.steps[0].should.have.property('task').equal('insert');
-      task.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      task.steps[0].should.have.property('percent').equal(100);
-      task.steps[0].should.have.property('lineRead').equal(20);
-      task.steps[0].should.have.property('took');
-      task.steps[0].should.have.property('status').equal('success');
+      expect(task.steps[0]).have.property('task').equal('insert');
+      expect(task.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(task.steps[0]).have.property('percent').equal(100);
+      expect(task.steps[0]).have.property('lineRead').equal(20);
+      expect(task.steps[0]).have.property('took');
+      expect(task.steps[0]).have.property('status').equal('success');
     });
 
     // test report
-    it('should get report with all informations', async () => {
+    it('Should get report with all informations from the insertion', async () => {
       const report = await getLatestReport();
 
-      report.should.have.property('done').equal(true);
-      report.should.have.property('currentTask').equal('end');
-      report.should.have.property('steps');
-      report.should.have.property('createdAt');
-      report.should.have.property('endAt');
-      report.should.have.property('took');
+      expect(report).have.property('done').equal(true);
+      expect(report).have.property('currentTask').equal('end');
+      expect(report).have.property('steps');
+      expect(report).have.property('createdAt');
+      expect(report).have.property('endAt');
+      expect(report).have.property('took');
 
-      report.steps[0].should.have.property('task').equal('insert');
-      report.steps[0].should.have.property('file').equal('fake1.jsonl.gz');
-      report.steps[0].should.have.property('percent').equal(100);
-      report.steps[0].should.have.property('lineRead').equal(20);
-      report.steps[0].should.have.property('took');
-      report.steps[0].should.have.property('status').equal('success');
+      expect(report.steps[0]).have.property('task').equal('insert');
+      expect(report.steps[0]).have.property('file').equal('fake1.jsonl.gz');
+      expect(report.steps[0]).have.property('percent').equal(100);
+      expect(report.steps[0]).have.property('lineRead').equal(20);
+      expect(report.steps[0]).have.property('took');
+      expect(report.steps[0]).have.property('status').equal('success');
     });
 
     after(async () => {
@@ -341,42 +319,42 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     });
   });
 
-  describe('/update/fake1.jsonl try to insert a file that is in the wrong format', () => {
+  describe('Don\'t do a insertion of a file already installed because the file is at the wrong format', () => {
     // test return message
-    it('should return a error message', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return a error message', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
 
-      response.should.have.status(400);
-      response.body.should.have.property('message').equal('name of file is in bad format (accepted a .gz file)');
+      expect(res).have.status(400);
+      expect(res.body.message).be.equal('name of file is in bad format (accepted a .gz file)');
     });
   });
 
-  describe('/update/fileDoesntExist.jsonl.gz try to insert a file that does not exist', () => {
+  describe('Don\'t do a insertion of a file already installed because the file doesn\'t exist on ezunpaywall', () => {
     // test return message
-    it('should return a error message', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return a error message', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fileDoesntExist.jsonl.gz')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
 
-      response.should.have.status(404);
-      response.body.should.have.property('message').equal('file doesn\'t exist');
+      expect(res).have.status(404);
+      expect(res.body.message).be.equal('file doesn\'t exist');
     });
   });
 
-  describe('/update/fake1.jsonl.gz?offset=100&limit=50 try to insert a file with limit < offset', () => {
+  describe('Don\'t do a insertion of a file already installed because the parameter limit can\t be lower than offset', () => {
     // test return message
-    it('should return a error message', async () => {
-      const response = await chai.request(ezunpaywallURL)
+    it('Should return a error message', async () => {
+      const res = await chai.request(ezunpaywallURL)
         .post('/update/fake1.jsonl.gz?offset=100&limit=50')
         .set('Access-Control-Allow-Origin', '*')
         .set('Content-Type', 'application/json');
 
-      response.should.have.status(400);
-      response.body.should.have.property('message').equal('limit can\t be lower than offset or 0');
+      expect(res).have.status(400);
+      expect(res.body.message).be.equal('limit can\t be lower than offset or 0');
     });
   });
 
@@ -384,7 +362,5 @@ describe('test insertion with a file already installed in ez-unpaywall', () => {
     await deleteIndex('unpaywall');
     await deleteIndex('task');
     await deleteFile('fake1.jsonl.gz');
-    await deleteFile('fake2.jsonl.gz');
-    await deleteFile('fake3.jsonl.gz');
   });
 });
